Concerns #25 - Use slow and fast instead of long and short
